+++ src/App.css
@@ -1,3 +1,9 @@
 .App {
   text-align: center;
 }
+
+
+.img-render {
+  object-fit: fit;
+  image-rendering: auto;
+}
